修复文件大小判断文件

+++ src/main/java/cn/jackding/aliststrm/service/CopyAlistFileService.java
@@ -70,7 +70,7 @@ public class CopyAlistFileService {
             } else {
                 //是视频文件才复制 并且不存在
                 if (!(200 == jsonObject.getInteger("code")) && Utils.isVideo(name)) {
-                    if (contentJson.getInteger("size") > Integer.parseInt(minFileSize) * 1024 * 1024) {
+                    if (contentJson.getLong("size") > Long.parseLong(minFileSize) * 1024 * 1024) {
                         alistService.copyAlist(srcDir + "/" + relativePath, dstDir + "/" + relativePath, Collections.singletonList(name));
                         flag.set(true);
                     }
